+++ packages/adminui-vue/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.5]
+### Changed
+ - !!Breaking!! `u-input` does not supports v-model, now gets and sets value from store automatically
+
+### Added
+ - `u-base-input` - when using type="number", helps to establish the precision of rounding and step
+
 ## [1.6.4]
 ### Fixed
  - "Always use domain for login" checkbox should be visible in navbar User menu in case "Negotiate" auth is enabled for app [UBDF-9586]